Implement and fix rgb and opacity
Did not fully read the assignment so I assumed
that the button for rgb was rainbow colors, not
random colors in rgb. Fixed it so now it will
be random colors every time the user hovers over
the grid.

Implemented the opacity button so that every time
the user hovers over the tiles, it gets darker
until it becomes completely black.